fuse: support buffer copying for kernel addresses
This is a preparatory patch needed to support pinned buffers in
fuse-over-io-uring. For pinned buffers, we get the vmapped address of
the buffer, which we can directly use with memcpy.

Currently, buffer copying in fuse only supports extracting underlying
pages from an iov iter and kmapping them. This commit allows buffer
copying to work directly on a kaddr.

Signed-off-by: Joanne Koong <joannelkoong@gmail.com>
Reviewed-by: Bernd Schubert <bernd@bsbernd.com>